feat: show selected resources in detail view
