+++ packages/web/app/src/components/organization/settings/access-tokens/create-access-token-sheet-content.tsx
@@ -25,6 +25,7 @@ import {
   type ResourceSelection,
 } from '../../members/resource-selector';
 import { SelectedPermissionOverview } from '../../members/selected-permission-overview';
+import { permissionLevelToResourceName, resolveResources } from './shared-helpers';
 
 /** @soure packages/services/api/src/modules/organization/providers/organization-access-tokens.ts */
 const TitleInputModel = z
@@ -408,88 +409,4 @@ function AcessTokenCreatedConfirmationDialogue(props: {
       </AlertDialog.AlertDialogContent>
     </AlertDialog.AlertDialog>
   );
-}
-
-function resolveResources(
-  organizationSlug: string,
-  resources: ResourceSelection,
-): null | Record<GraphQLSchema.PermissionLevel, Array<string>> {
-  if (resources.mode === GraphQLSchema.ResourceAssignmentMode.All || !resources.projects) {
-    return null;
-  }
-
-  const resolvedResources: Record<GraphQLSchema.PermissionLevel, Array<string>> = {
-    [GraphQLSchema.PermissionLevel.Organization]: [organizationSlug],
-    [GraphQLSchema.PermissionLevel.Project]: [],
-    [GraphQLSchema.PermissionLevel.Target]: [],
-    [GraphQLSchema.PermissionLevel.AppDeployment]: [],
-    [GraphQLSchema.PermissionLevel.Service]: [],
-  };
-
-  for (const project of resources.projects) {
-    resolvedResources[GraphQLSchema.PermissionLevel.Project].push(
-      `${organizationSlug}/${project.projectSlug}`,
-    );
-    if (project.targets.mode === GraphQLSchema.ResourceAssignmentMode.All) {
-      resolvedResources[GraphQLSchema.PermissionLevel.Target].push(
-        `${organizationSlug}/${project.projectSlug}/*`,
-      );
-      resolvedResources[GraphQLSchema.PermissionLevel.Service].push(
-        `${organizationSlug}/${project.projectSlug}/*/service/*`,
-      );
-      resolvedResources[GraphQLSchema.PermissionLevel.AppDeployment].push(
-        `${organizationSlug}/${project.projectSlug}/*/appDeployment/*`,
-      );
-      continue;
-    }
-    for (const target of project.targets.targets) {
-      resolvedResources[GraphQLSchema.PermissionLevel.Target].push(
-        `${organizationSlug}/${project.projectSlug}/${target.targetSlug}`,
-      );
-      if (target.services.mode === GraphQLSchema.ResourceAssignmentMode.All) {
-        resolvedResources[GraphQLSchema.PermissionLevel.Service].push(
-          `${organizationSlug}/${project.projectSlug}/${target.targetSlug}/service/*`,
-        );
-      } else if (target.services.services) {
-        for (const service of target.services.services) {
-          resolvedResources[GraphQLSchema.PermissionLevel.Service].push(
-            `${organizationSlug}/${project.projectSlug}/${target.targetSlug}/service/${service.serviceName}`,
-          );
-        }
-      }
-      if (target.appDeployments.mode === GraphQLSchema.ResourceAssignmentMode.All) {
-        resolvedResources[GraphQLSchema.PermissionLevel.AppDeployment].push(
-          `${organizationSlug}/${project.projectSlug}/${target.targetSlug}/appDeployment/*`,
-        );
-      } else if (target.appDeployments.appDeployments) {
-        for (const appDeployment of target.appDeployments.appDeployments) {
-          resolvedResources[GraphQLSchema.PermissionLevel.AppDeployment].push(
-            `${organizationSlug}/${project.projectSlug}/${target.targetSlug}/appDeployment/${appDeployment.appDeployment}`,
-          );
-        }
-      }
-    }
-  }
-
-  return resolvedResources;
-}
-
-function permissionLevelToResourceName(level: GraphQLSchema.PermissionLevel) {
-  switch (level) {
-    case GraphQLSchema.PermissionLevel.Organization: {
-      return 'organizations';
-    }
-    case GraphQLSchema.PermissionLevel.Project: {
-      return 'projects';
-    }
-    case GraphQLSchema.PermissionLevel.Target: {
-      return 'targets';
-    }
-    case GraphQLSchema.PermissionLevel.Service: {
-      return 'services';
-    }
-    case GraphQLSchema.PermissionLevel.AppDeployment: {
-      return 'app deployments';
-    }
-  }
 }
